+++ src/gui/GuiCommon.py
@@ -418,6 +418,24 @@ class SwitchButton():
         self.btn.pack(*args, **kwargs)
 
     # ===============================================================================
+    # @brief:   Override switch to OFF state
+    #
+    # @return:      void
+    # ===============================================================================  
+    def turn_off(self):
+        self.state = False
+        self.__update_appear(self.state)
+
+    # ===============================================================================
+    # @brief:   Override switch to ON state
+    #
+    # @return:      void
+    # ===============================================================================  
+    def turn_on(self):
+        self.state = True
+        self.__update_appear(self.state)
+
+    # ===============================================================================
     # @brief:   Refresh appearance
     #
     # @param[in]:   state   - State of toggle switch